requirements.txt
Django==1.10.2
djangorestframework==3.4.7
multidict==2.0.1
python-json-logger==0.1.5
PyYAML==3.12
